requirements.txt
supabase
feedparser
openai
beautifulsoup4
lxml
fastapi
uvicorn
pydantic
python-dotenv
requests
duckduckgo_search
googlesearch-python
markdown
youtube-transcript-api
polar-sdk
